pyproject.toml
[project]
name = ""
version = ""
description = ""
authors = [
    {name = "Mateus Lopes Albano", email = "mateusalbano22@gmail.com"},
]
dependencies = [
    "django>=4.2.5",
    "djangorestframework>=3.14.0",
    "django-cors-headers>=4.2.0",
    "djangorestframework-simplejwt>=5.3.0",
    "setuptools>=68.2.2",
    "python-magic>=0.4.27",
    "Pillow>=10.0.1",
]
requires-python = ">=3.10"
readme = "README.md"
license = {text = "MIT"}

[tool.pdm.scripts]
dev = "python manage.py runserver"
runserver = "python manage.py runserver"
createsuperuser = "python manage.py createsuperuser"
migrate = "python manage.py migrate"
makemigrations = "python manage.py makemigrations"
shell = "python manage.py shell"
test = "python manage.py test"
startapp = "python manage.py startapp {args}"
loaddata = "python manage.py loaddata {args}"
dumpdata = "python manage.py dumpdata {args}"

[tool.pdm.dev-dependencies]
dev = [
    "isort>=5.12.0",
    "black>=23.9.1",
]

[tool.black]
line-length = 120

[[tool.pdm.autoexport]]
filename = "requirements.txt"
without-hashes = "true"
